+++ ChangeLog.md
@@ -6,6 +6,13 @@ Planned items for Conductor 1.1.0
 * Common install script behaviors/wrappers
 * Refinements to converter, as needed
 
+Conductor 1.0.1
+---------------
+
+* Added an autoloader.
+* Made it easier to run a PEAR Package conversion to Composer via
+  command-line tools and a config file.
+  
 
 Conductor 1.0.0
 ---------------